requirements.txt
playwright==1.54.0
playwright-stealth==2.0.3
rich==13.9.4
python-dotenv==1.0.1
openai>=2.31.0
langgraph>=1.1.6
langchain-core>=1.2.28
